+++ lib/assignable_values/errors.rb
@@ -3,4 +3,5 @@ module AssignableValues
   class DelegateUnavailable < Error; end
   class NoValuesGiven < Error; end
   class NoDefault < Error; end
+  class UnsupportedOption < Error; end
 end